Only copy config files to /etc if they don't exist
This was clobbering my local testing changes, if I want clean configs
I'll delete and re-provision.

+++ Vagrantfile
@@ -64,8 +64,15 @@ Vagrant::Config.run("2") do |config|
 
       # Copy configs in place
       mkdir -p /etc/pantry
-      cp /vagrant/dist/client.yml /etc/pantry/client.yml
-      cp /vagrant/dist/server.yml /etc/pantry/server.yml
+
+      if [ ! -f /etc/pantry/client.yml ]; then
+        cp /vagrant/dist/client.yml /etc/pantry/client.yml
+      fi
+
+      if [ ! -f /etc/pantry/server.yml ]; then
+        cp /vagrant/dist/server.yml /etc/pantry/server.yml
+      fi
+
       cp /vagrant/dist/upstart/pantry-client.conf /etc/init/pantry-client.conf
       cp /vagrant/dist/upstart/pantry-server.conf /etc/init/pantry-server.conf
